refactor: improve code organization gmail, register and user service
