--- src/main/java/com/tech_symfony/movie_booking/api/user/Provider.java
@@ -1,5 +1,0 @@
-package com.tech_symfony.movie_booking.api.user;
-
-public enum Provider {
-	LOCAL, GOOGLE
-}
