--- src/main/java/com/tech_symfony/movie_booking/api/user/UnverifiedUserException.java
@@ -1,8 +1,0 @@
-package com.tech_symfony.movie_booking.api.user;
-
-
-public class UnverifiedUserException extends RuntimeException {
-    public UnverifiedUserException() {
-        super("User is not verified");
-    }
-}
