+++ src/main/java/com/tech_symfony/movie_booking/api/user/attribute/Provider.java
@@ -1,0 +1,5 @@
+package com.tech_symfony.movie_booking.api.user.attribute;
+
+public enum Provider {
+	LOCAL, GOOGLE
+}
