+++ src/main/java/com/tech_symfony/movie_booking/api/user/auth/CustomUserDetailService.java
@@ -1,5 +1,7 @@
-package com.tech_symfony.movie_booking.api.user;
+package com.tech_symfony.movie_booking.api.user.auth;
 
+import com.tech_symfony.movie_booking.api.user.User;
+import com.tech_symfony.movie_booking.api.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
