+++ src/main/java/com/tech_symfony/movie_booking/api/user/register/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.tech_symfony.movie_booking.api.user;
+package com.tech_symfony.movie_booking.api.user.register;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
