+++ src/main/java/com/tech_symfony/movie_booking/api/user/register/UserRegisterController.java
@@ -1,23 +1,23 @@
-package com.tech_symfony.movie_booking.api.user;
+package com.tech_symfony.movie_booking.api.user.register;
 
+import com.tech_symfony.movie_booking.api.user.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.rest.webmvc.BasePathAwareController;
-import org.springframework.data.rest.webmvc.RepositoryRestController;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @BasePathAwareController
 @RequiredArgsConstructor
-public class UserController {
+public class UserRegisterController {
 
 	private final UserService userService;
 
 	@PostMapping("/auth/register")
 	@ResponseBody
 	public String register(@RequestBody @Valid RegisterRequest request) {
-        return userService.register(request);
+		return userService.register(request);
 	}
 
 	@GetMapping("/auth/verify")
