--- src/main/java/com/tech_symfony/movie_booking/system/config/ServletUriConfig.java
@@ -1,5 +1,0 @@
-package com.tech_symfony.movie_booking.system.config;
-
-public class ServletUriConfig {
-	
-}
