+++ src/main/java/com/tech_symfony/movie_booking/system/config/api/ServletUriConfig.java
@@ -1,0 +1,5 @@
+package com.tech_symfony.movie_booking.system.config.api;
+
+public class ServletUriConfig {
+
+}
